// src/sync-client/model/chunk.ts
/**
 * Anki Sync Api
 * Api for Anki Synchronization
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Chunk { 
    done: boolean;
    notes: Array<Array<number | string>>;
    cards: Array<Array<number | string>>;
    revlog: Array<Array<number>>;
}