// src/sync-client/model/note.ts
/**
 * Anki Sync Api
 * Api for Anki Synchronization
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Note { 
    id: number;
    guid: number;
    mid: number;
    mod: number;
    usn: number;
    tags: number;
    flds: string;
    sfld: number;
    csum: number;
    flags: number;
    data: string;
}